requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-aarch64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=51_gnu
affine=2.4.0=pyhd8ed1ab_1
annotated-types=0.7.0=pyhd8ed1ab_1
apache-sedona=1.7.1=py310h78583b1_0
arrow-cpp=19.0.0=h0040ae0_3
attrs=25.3.0=pyh71513ae_0
aws-c-auth=0.8.5=h998d150_0
aws-c-cal=0.8.5=h6ac735f_0
aws-c-common=0.11.1=h998d150_0
aws-c-compression=0.3.1=h998d150_0
aws-c-event-stream=0.5.4=h419075a_0
aws-c-http=0.9.3=h998d150_0
aws-c-io=0.17.0=h998d150_0
aws-c-mqtt=0.12.2=h998d150_0
aws-c-s3=0.7.11=h6ac735f_0
aws-c-sdkutils=0.2.3=h998d150_0
aws-checksums=0.2.3=h998d150_0
aws-crt-cpp=0.31.0=h419075a_0
aws-glue-libs=4.0.0=pypi_0
aws-sdk-cpp=1.11.528=hc0ddfc7_0
blas=2.131=openblas
blas-devel=3.9.0=31_h9678261_openblas
blosc=1.21.3=h419075a_0
boost-cpp=1.82.0=hb8fdbf2_2
boto3=1.39.11=pyhd8ed1ab_0
botocore=1.39.12=pyge310_1234567_0
bottleneck=1.4.2=py310hf6ef57e_0
brotli-python=1.0.9=py310he22ea3a_6
bzip2=1.0.8=h998d150_6
c-ares=1.19.1=h998d150_0
ca-certificates=2025.7.14=hbd8a1cb_0
cairo=1.16.0=hd83b346_6
certifi=2025.7.14=pyhd8ed1ab_0
cffi=1.17.1=py310h1451162_0
cfitsio=3.470=h152aa4d_7
click=8.2.1=pyh707e725_0
click-plugins=1.1.1.2=pyhd8ed1ab_0
cligj=0.7.2=pyhd8ed1ab_2
cyrus-sasl=2.1.28=h6c5dea3_0
expat=2.7.1=h419075a_0
fontconfig=2.14.1=h881afe3_3
freetype=2.13.3=h6df46f4_0
freexl=2.0.0=h5428426_0
geos=3.10.6=h419075a_0
geotiff=1.7.0=h975d5ae_4
gflags=2.2.2=h54f1f3f_1004
giflib=5.2.2=h31becfc_0
glog=0.5.0=hbffda88_0
h2=4.2.0=pyhd8ed1ab_0
h3-py=3.7.4=py310h16c9dca_0
h3-pyspark=1.2.6=pyhd8ed1ab_0
hdf4=4.2.13=h96bad59_4
hdf5=1.14.5=h2117f30_2
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.1=h419075a_0
jmespath=1.0.1=pyhd8ed1ab_1
jpeg=9e=h2a766a3_3
json-c=0.16=hcc5b78b_0
kealib=1.5.0=h1f8070d_2
krb5=1.21.3=ha46f760_1
lcms2=2.16=had1b715_1
ld_impl_linux-aarch64=2.40=h48e3ba3_0
lerc=4.0.0=hfdc4d58_1
libabseil=20250127.0=cxx17_h419075a_0
libblas=3.9.0=31_h1a9f1db_openblas
libboost=1.82.0=hda0696e_2
libbrotlicommon=1.0.9=h4e544f5_7
libbrotlidec=1.0.9=h4e544f5_7
libbrotlienc=1.0.9=h4e544f5_7
libcblas=3.9.0=31_hab92f65_openblas
libcurl=8.14.1=h2e08ae6_0
libdeflate=1.22=h86ecc28_0
libedit=3.1.20230828=h998d150_0
libev=4.33=h516909a_1
libevent=2.1.12=h6ac735f_1
libexpat=2.7.1=hfae3067_0
libffi=3.4.4=h419075a_1
libgcc=15.1.0=he277a41_3
libgcc-ng=15.1.0=he9431aa_3
libgdal=3.6.2=h9fd4a1d_10
libgfortran-ng=13.2.0=he9431aa_0
libgfortran5=13.2.0=h582850c_0
libglib=2.84.2=h4c1cdc3_0
libgomp=15.1.0=he277a41_3
libgrpc=1.71.0=hb788212_0
libiconv=1.18=hc99b53d_1
libkml=1.3.0=hadc4260_7
liblapack=3.9.0=31_h411afd4_openblas
liblapacke=3.9.0=31_hc659ca5_openblas
libnetcdf=4.8.1=h5c2efc6_5
libnghttp2=1.57.0=hb788212_0
libntlm=1.4=hf897c2e_1002
libopenblas=0.3.29=h29fea54_0
libpng=1.6.39=h998d150_0
libpq=17.4=h33f8927_1
libprotobuf=5.29.3=h708d2a9_1
libre2-11=2024.07.02=h419075a_0
libspatialite=5.1.0=ha2ba0a8_2
libssh2=1.11.1=hfa2bbb0_0
libstdcxx=15.1.0=h3f4de04_3
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=hb2e9abc_2
libtiff=4.7.0=heeacae9_0
libuuid=1.41.5=h998d150_0
libwebp-base=1.6.0=ha2e29f5_0
libxcb=1.17.0=hf66535e_0
libxcrypt=4.4.36=h31becfc_1
libxml2=2.13.8=h6097fa9_0
libzip=1.8.0=h3b934f4_1
lz4-c=1.9.4=h419075a_1
minizip=4.0.3=ha60d17d_0
ncurses=6.4=h419075a_0
nomkl=3.0=0
nspr=4.36=h5ad3122_0
nss=3.89.1=h419075a_0
numexpr=2.11.0=py310hb87de48_0
numpy=1.26.4=py310he45c16d_0
numpy-base=1.26.4=py310h15d264d_0
openblas=0.3.29=hd43f75c_0
openblas-devel=0.3.29=h29fea54_0
openjpeg=2.5.2=ha0e65c7_1
openldap=2.6.10=h30c48ee_0
openssl=3.5.1=hd08dc88_0
orc=2.1.1=hdd9cb92_0
pandas=2.2.3=py310h419075a_0
pcre2=10.42=hcfaa891_1
pip=25.1=pyhc872135_2
pixman=0.46.2=h86a87f0_0
poppler=24.09.0=h2917649_3
poppler-data=0.4.12=hd8ed1ab_0
postgresql=17.4=hebb359e_1
proj=9.3.1=he46dc16_1
pthread-stubs=0.3=hfd63f10_1
py4j=0.10.9.5=pyhd8ed1ab_0
pyarrow=19.0.0=py310h419075a_1
pycparser=2.22=pyh29332c3_1
pydantic=2.11.7=pyh3cfb1c2_0
pydantic-core=2.33.2=py310hccdbd67_0
pyparsing=3.2.3=pyhd8ed1ab_1
pysocks=1.7.1=pyha55dd90_7
pyspark=3.3.0=pyhd8ed1ab_0
python=3.10.18=hbb0f47a_0
python-dateutil=2.9.0.post0=pyhe01879c_2
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0=py310h7cee911_3
qhull=2020.2=hd62202e_2
rasterio=1.4.3=py310h66c4f06_0
re2=2024.07.02=hb8fdbf2_0
readline=8.2=h998d150_0
s2n=1.5.14=h6ac735f_0
s3transfer=0.13.1=pyhd8ed1ab_0
setuptools=78.1.1=py310hd43f75c_0
shapely=2.0.6=py310hd888f1e_0
six=1.17.0=pyhd8ed1ab_0
snappy=1.2.1=h419075a_0
sqlite=3.50.2=h998d150_1
tiledb=2.3.3=h3849020_3
tk=8.6.14=hb5ae6a8_1
typing-extensions=4.14.1=h4440ef1_0
typing-inspection=0.4.1=pyhd8ed1ab_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h04d1e81_0
uriparser=0.9.8=ha6cfe40_0
urllib3=2.5.0=pyhd8ed1ab_0
utf8proc=2.6.1=h998d150_1
wheel=0.45.1=py310hd43f75c_0
xerces-c=3.2.4=h419075a_1
xorg-libx11=1.8.12=hf66535e_1
xorg-libxau=1.0.12=hf66535e_0
xorg-libxdmcp=1.1.5=hf66535e_0
xorg-libxext=1.3.6=h57736b2_0
xorg-libxrender=0.9.12=h86ecc28_0
xorg-xorgproto=2024.1=h998d150_1
xz=5.6.4=h998d150_1
yaml=0.2.5=hf897c2e_2
zlib=1.2.13=h998d150_1
zstandard=0.23.0=py310hc476304_1
zstd=1.5.6=h6a09583_0
